update activity step order from recipe if count==1
Activities can have multiple recipes, and in that case, the admin user
must impose a strict ordering of all steps. However, in the case where
there is one recipe, it is more convenient to just automatically use the
order of the steps within the recipe for the activity step order.