use object for storing views
Keep a reference from the option uid to the option view so that we can
do a look up of the view based on uid. We needed this to allow us to
serialize the options in the correct order when saving (we were
serializing in the array order before, which didn't account for changes
in option ordering).