Update the GalleryController#index_as_json endpoint to call through to Algolia.
This is legacy only for the cs-mobile app until we change it to call Algolia directly
There are no other clients.
We'd have to do this anyhow b/c even if we push an app update, you'll have old versions out there that depend on this API for a long time.
This is a lot of duplication of what is in api.js.coffee, not delightful.