Never proxy API!
For activities, running a split test where those whose slug starts with a-m just returns
the normal page to google, and n-z returns the snapshot. Right now only for google as we
don't think other crawlers are smart enough to run the javascript yet. The hope is that this
will help with including images in SERPs.